Remove duplicated `ShowkaseButton` and fix colors.
